+++ src/main/java/com/telepathicgrunt/the_bumblezone/items/StingerSpearItem.java
@@ -2,6 +2,8 @@ package com.telepathicgrunt.the_bumblezone.items;
 
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
+import com.telepathicgrunt.the_bumblezone.entities.nonliving.ThrownStingerSpearEntity;
+import com.telepathicgrunt.the_bumblezone.tags.BzItemTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
@@ -10,6 +12,8 @@ import net.minecraft.stats.Stats;
 import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.MoverType;
@@ -21,6 +25,7 @@ import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.entity.projectile.ThrownTrident;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TridentItem;
 import net.minecraft.world.item.UseAnim;
 import net.minecraft.world.item.Vanishable;
 import net.minecraft.world.item.enchantment.Enchantment;
@@ -30,31 +35,30 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.Vec3;
 
-public class StingerSpearItem extends Item implements Vanishable {
-
+public class StingerSpearItem extends TridentItem {
+    public static final float BASE_DAMAGE = 1F;
+    public static final float BASE_THROWN_DAMAGE = 1.5F;
     private final Multimap<Attribute, AttributeModifier> defaultModifiers;
 
     public StingerSpearItem(Properties properties) {
         super(properties.durability(220));
         ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
-        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", 3.0, AttributeModifier.Operation.ADDITION));
+        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", BASE_DAMAGE, AttributeModifier.Operation.ADDITION));
         builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Tool modifier", -1F, AttributeModifier.Operation.ADDITION));
         this.defaultModifiers = builder.build();
     }
 
     @Override
-    public boolean canAttackBlock(BlockState blockState, Level level, BlockPos blockPos, Player player) {
-        return !player.isCreative();
-    }
-
-    @Override
-    public UseAnim getUseAnimation(ItemStack itemStack) {
-        return UseAnim.SPEAR;
-    }
-
-    @Override
     public int getUseDuration(ItemStack itemStack) {
         return 50000;
+    }
+
+    /**
+     * Specify what item can repair this weapon
+     */
+    @Override
+    public boolean isValidRepairItem(ItemStack toRepair, ItemStack repair) {
+        return BzItemTags.STINGER_SPEAR_REPAIR_ITEMS.contains(repair.getItem());
     }
 
     @Override
@@ -64,14 +68,14 @@ public class StingerSpearItem extends Item implements Vanishable {
             if (remainingDuration >= 10) {
                 if (!level.isClientSide) {
                     itemStack.hurtAndBreak(1, player, playerx -> playerx.broadcastBreakEvent(livingEntity.getUsedItemHand()));
-                    ThrownTrident thrownTrident = new ThrownTrident(level, player, itemStack);
-                    thrownTrident.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, 2.5F + 0.5F, 1.0F);
+                    ThrownStingerSpearEntity thrownStingerSpear = new ThrownStingerSpearEntity(level, player, itemStack);
+                    thrownStingerSpear.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, 2.5F + 0.5F, 1.0F);
                     if (player.getAbilities().instabuild) {
-                        thrownTrident.pickup = AbstractArrow.Pickup.CREATIVE_ONLY;
+                        thrownStingerSpear.pickup = AbstractArrow.Pickup.CREATIVE_ONLY;
                     }
 
-                    level.addFreshEntity(thrownTrident);
-                    level.playSound(null, thrownTrident, SoundEvents.TRIDENT_THROW, SoundSource.PLAYERS, 1.0F, 1.0F);
+                    level.addFreshEntity(thrownStingerSpear);
+                    level.playSound(null, thrownStingerSpear, SoundEvents.TRIDENT_THROW, SoundSource.PLAYERS, 1.0F, 1.0F);
                     if (!player.getAbilities().instabuild) {
                         player.getInventory().removeItem(itemStack);
                     }
@@ -95,23 +99,21 @@ public class StingerSpearItem extends Item implements Vanishable {
     }
 
     @Override
-    public boolean hurtEnemy(ItemStack itemStack, LivingEntity livingEntity, LivingEntity livingEntity2) {
-        itemStack.hurtAndBreak(1, livingEntity2, livingEntityx -> livingEntityx.broadcastBreakEvent(EquipmentSlot.MAINHAND));
-        return true;
-    }
-
-    @Override
-    public boolean mineBlock(ItemStack itemStack, Level level, BlockState blockState, BlockPos blockPos, LivingEntity livingEntity) {
-        if ((double)blockState.getDestroySpeed(level, blockPos) != 0.0) {
-            itemStack.hurtAndBreak(5, livingEntity, livingEntityx -> livingEntityx.broadcastBreakEvent(EquipmentSlot.MAINHAND));
-        }
-
-        return true;
-    }
-
-    @Override
     public Multimap<Attribute, AttributeModifier> getDefaultAttributeModifiers(EquipmentSlot equipmentSlot) {
         return equipmentSlot == EquipmentSlot.MAINHAND ? this.defaultModifiers : super.getDefaultAttributeModifiers(equipmentSlot);
+    }
+
+    @Override
+    public boolean hurtEnemy(ItemStack itemStack, LivingEntity enemy, LivingEntity user) {
+        enemy.addEffect(new MobEffectInstance(
+                MobEffects.POISON,
+                200,
+                0,
+                true,
+                true,
+                true));
+        itemStack.hurtAndBreak(1, user, (entity) -> entity.broadcastBreakEvent(EquipmentSlot.MAINHAND));
+        return true;
     }
 
     /**
@@ -124,10 +126,5 @@ public class StingerSpearItem extends Item implements Vanishable {
         }
 
         return enchantment.category.canEnchant(stack.getItem());
-    }
-
-    @Override
-    public int getEnchantmentValue() {
-        return 1;
     }
 }